% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_up_docker.R
\name{clean_up_docker}
\alias{clean_up_docker}
\title{Cleans up docker container}
\usage{
clean_up_docker(
  container_name = "test",
  parameters_monkeys = parameters_monkeys
)
}
\arguments{
\item{container_name}{}

\item{parameters_monkeys}{}
}
\description{
Cleans up docker container
}
