% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_task.R
\name{complete_task}
\alias{complete_task}
\title{Completes a jsPsychMaker task}
\usage{
complete_task(parameters_monkeys, uid, links, remote_driver = NULL)
}
\arguments{
\item{parameters_monkeys}{The parameters_monkeys list}

\item{uid}{User id}

\item{links}{Links to complete}

\item{remote_driver}{remote_driver output}
}
\value{
Completes a task and returns a container name.
}
\description{
Completes a jsPsychMaker task
}
