% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{copy_files_to_data}
\alias{copy_files_to_data}
\title{In local protocols, move the Downloaded csv to the data folder of the protocol}
\usage{
copy_files_to_data(pre_existing_CSV, parameters_monkeys, uid, task = task)
}
\arguments{
\item{pre_existing_CSV}{vector with filenames from a previous call of check_Downloads()}

\item{parameters_monkeys}{parameters_monkeys list}

\item{uid}{User id}

\item{task}{Last task in _targets pipeline}
}
\value{
copies the new csv files to .data
}
\description{
In local protocols, move the Downloaded csv to the data folder of the protocol
}
