% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_docker.R
\name{create_docker}
\alias{create_docker}
\title{Creates a docker container}
\usage{
create_docker(uid = "test", parameters_monkeys = NULL)
}
\arguments{
\item{uid}{user id}

\item{parameters_monkeys}{parameters_monkeys list of parameters}
}
\value{
A list with the parameters of the newly created docker container
}
\description{
Creates a docker container
}
