% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_links.R
\name{create_links}
\alias{create_links}
\title{create_links}
\usage{
create_links(parameters_monkeys, uid, remote_driver = NULL)
}
\arguments{
\item{parameters_monkeys}{parameters_monkeys list}

\item{uid}{user id}

\item{remote_driver}{remDr}
}
\value{
A list of links
}
\description{
create_links
}
