% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_remDr.R
\name{create_remDr}
\alias{create_remDr}
\title{Creates remote driver in docker container (opens browser)}
\usage{
create_remDr(
  container_port,
  browserName,
  container_name = NULL,
  parameters_monkeys = parameters_monkeys
)
}
\arguments{
\item{container_port}{port of container}

\item{browserName}{c("chrome", "firefox")}

\item{container_name}{name of container}

\item{parameters_monkeys}{parameters_monkeys list of parameters}
}
\value{
A remDr object
}
\description{
Creates remote driver in docker container (opens browser)
}
