% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{create_targets_file}
\alias{create_targets_file}
\title{Create _targets.R file from a protocol folder}
\usage{
create_targets_file(
  folder = "~/Downloads/",
  credentials_folder = ".vault/",
  uid = 1,
  times_repeat_protocol = 1,
  time_to_sleep_before_repeating_protocol = 1,
  browserName = "chrome",
  big_container = FALSE,
  keep_alive = FALSE,
  folder_downloads = NULL,
  DEBUG = FALSE,
  screenshot = FALSE,
  debug_file = FALSE,
  console_logs = FALSE,
  open_VNC = FALSE,
  pid = NULL,
  uid_URL = TRUE,
  local_or_server = NULL,
  local_folder_tasks = NULL,
  server_folder_tasks = NULL,
  disable_web_security = FALSE,
  initial_wait = 2,
  wait_retry = 0.1,
  forced_random_wait = FALSE,
  forced_refresh = NULL,
  forced_seed = NULL,
  dont_ask = FALSE
)
}
\arguments{
\item{folder}{protocol folder}

\item{credentials_folder}{folder where files "SERVER_PATH.R" and ".credentials" are. Usually .vault/}

\item{uid}{.}

\item{times_repeat_protocol}{if different than 1, creates an URL parameter ID and multiple links changing the uid}

\item{time_to_sleep_before_repeating_protocol}{In seconds, how long to sleep before repeating protocol}

\item{browserName}{.}

\item{big_container}{.}

\item{keep_alive}{.}

\item{folder_downloads}{.}

\item{DEBUG}{.}

\item{screenshot}{.}

\item{debug_file}{.}

\item{console_logs}{.}

\item{open_VNC}{.}

\item{pid}{project id}

\item{uid_URL}{.}

\item{local_or_server}{.}

\item{local_folder_tasks}{.}

\item{server_folder_tasks}{.}

\item{disable_web_security}{.}

\item{initial_wait}{.}

\item{wait_retry}{.}

\item{forced_random_wait}{.}

\item{forced_refresh}{.}

\item{forced_seed}{.}

\item{dont_ask}{do everything without asking for user input}
}
\value{
Creates a _targets.R file adding parameters_monkeys_minimal
}
\description{
Create _targets.R file from a protocol folder
}
