% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_credentials.R
\name{get_credentials}
\alias{get_credentials}
\title{Get credentials from folder}
\usage{
get_credentials(credentials_folder)
}
\arguments{
\item{credentials_folder}{folder where files "SERVER_PATH.R" and ".credentials" are. Usually .vault/}
}
\value{
A list with credentials$value and server_path$value
}
\description{
Get credentials from folder
}
