% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{list_data_server}
\alias{list_data_server}
\title{List files in the .data folder in the server}
\usage{
list_data_server(pid, credentials_file = NULL)
}
\arguments{
\item{pid}{protocol id}

\item{credentials_file}{Path to .credentials file. Usually: \code{.vault/.credentials}}
}
\value{
A tibble with a list of files
}
\description{
List files in the .data folder in the server
}
