% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{parse_elements}
\alias{parse_elements}
\title{Parse website elements to extract and tabulate all the properties}
\usage{
parse_elements(what, page_source_rvest)
}
\arguments{
\item{what}{a html tag such as c("p", "textarea", "input", "div", ...)}

\item{page_source_rvest}{page source rvest object}
}
\value{
A DF
}
\description{
Parse website elements to extract and tabulate all the properties
}
