% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{random_number_trimmed}
\alias{random_number_trimmed}
\title{Generate random numbers from a trimmed normal distribution}
\usage{
random_number_trimmed(
  n = 1,
  min = 0,
  max = 100,
  mean = NULL,
  sd = NULL,
  round = TRUE,
  distrib = "normal"
)
}
\arguments{
\item{n}{number of random numbers to generate}

\item{min}{min value}

\item{max}{max value}

\item{mean}{mean of normal distribution}

\item{sd}{standard deviation of normal distribution}

\item{round}{TRUE / FALSE}

\item{distrib}{One of c("normal", "uniform")}
}
\value{
Random numbers from a trimmed normal distribution
}
\description{
Generate random numbers from a trimmed normal distribution
}
\examples{
hist(random_number_trimmed(n = 1000, mean = 80))
}
