% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{reconnect_to_VNC}
\alias{reconnect_to_VNC}
\title{reconnect_to_VNC}
\usage{
reconnect_to_VNC(
  container_name = NULL,
  just_check = FALSE,
  port = NULL,
  DEBUG = FALSE
)
}
\arguments{
\item{container_name}{if empty, show available containers}

\item{port}{No need to specify one}

\item{DEBUG}{TRUE/FALSE}
}
\description{
reconnect_to_VNC
}
\examples{
reconnect_to_VNC()
}
