% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_select_input.R
\name{select_input}
\alias{select_input}
\title{Helper to select and interact with input fields}
\usage{
select_input(list_get_elements, remDr = NULL, DEBUG = FALSE, seed = 1)
}
\arguments{
\item{list_get_elements}{list_get_elements}

\item{remDr}{.}

\item{DEBUG}{TRUE / FALSE}

\item{seed}{numeric random seed}
}
\value{
list with selected input and the text of the selected element
}
\description{
Helper to select and interact with input fields
}
