% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{setup_folders}
\alias{setup_folders}
\title{setup_folders
Extract jsPsychMonkeys.zip and make sure we have all necessary folders in a specific location}
\usage{
setup_folders(folder, extract_zip = FALSE)
}
\arguments{
\item{folder}{destination folder for the project}

\item{extract_zip}{If TRUE, extracts jsPsychMonkeys.zip to folder}
}
\description{
setup_folders
Extract jsPsychMonkeys.zip and make sure we have all necessary folders in a specific location
}
\examples{
setup_folders(tempdir())
}
