% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{check_accept_alert}
\alias{check_accept_alert}
\title{check_accept_alert
Safely checks if there in an alert and accepts it}
\usage{
check_accept_alert(wait_retry = 0.5, remDr, DEBUG)
}
\description{
check_accept_alert
Safely checks if there in an alert and accepts it
}
