% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{check_Downloads}
\alias{check_Downloads}
\title{Check csv files in a folder}
\usage{
check_Downloads(parameters_monkeys, uid = "", links_tar = "")
}
\arguments{
\item{parameters_monkeys}{parameter_monkeys list}

\item{uid}{user id}

\item{links_tar}{targets object dependency to control when this target runs}
}
\value{
A list of csv files
}
\description{
Check csv files in a folder
}
