% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{check_trialids}
\alias{check_trialids}
\title{check_trialids}
\usage{
check_trialids(local_folder_protocol)
}
\arguments{
\item{local_folder_tasks}{local folders where the tasks are}
}
\description{
Checks that trialid's of an experiment in a folder follow the standard expected rules
}
