% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{copy_files_to_data}
\alias{copy_files_to_data}
\title{In local protocols, move the Downloaded csv to the data folder of the protocol}
\usage{
copy_files_to_data(pre_existing_CSV, parameters_monkeys, uid, task = task)
}
\arguments{
\item{task}{}
}
\description{
In local protocols, move the Downloaded csv to the data folder of the protocol
}
