% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{create_monkeys_project}
\alias{create_monkeys_project}
\title{Create a jsPsychMonkeys project for your data}
\usage{
create_monkeys_project(
  folder = "~/Downloads/",
  uid = 1,
  browserName = "chrome",
  big_container = FALSE,
  keep_alive = FALSE,
  folder_downloads = NULL,
  DEBUG = FALSE,
  screenshot = FALSE,
  debug_file = FALSE,
  console_logs = FALSE,
  open_VNC = FALSE,
  pid = NULL,
  uid_URL = TRUE,
  local_or_server = NULL,
  local_folder_tasks = NULL,
  server_folder_tasks = NULL,
  disable_web_security = FALSE,
  initial_wait = 2,
  wait_retry = 0.1,
  forced_random_wait = FALSE,
  forced_refresh = NULL,
  forced_seed = NULL,
  dont_ask = FALSE,
  open_rstudio = TRUE
)
}
\arguments{
\item{folder}{location for the project}

\item{pid}{project id}

\item{dont_ask}{answer YES to all questions so the process runs uninterrupted. This will:}

\item{open_rstudio}{Open RStudio with the new project TRUE / FALSE}
}
\value{
Opens a new RStudio project
}
\description{
run_initial_setup() will read your data and create a jsPsychMonkeys project
tailoring the _targets.R file to the tasks included in the data.
}
\examples{
run_initial_setup(pid = 999, download_files = FALSE,
data_location = system.file("extdata", package = "jsPsychMonkeys"),
download_task_script = FALSE,
folder = tempdir(),
sensitive_tasks = c(""), dont_ask = TRUE, open_rstudio = FALSE)
}
