% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{create_targets_file}
\alias{create_targets_file}
\title{Create _targets.R file from a protocol folder}
\usage{
create_targets_file(
  folder = "~/Downloads/",
  uid = 1,
  browserName = "chrome",
  big_container = FALSE,
  keep_alive = FALSE,
  folder_downloads = NULL,
  DEBUG = FALSE,
  screenshot = FALSE,
  debug_file = FALSE,
  console_logs = FALSE,
  open_VNC = FALSE,
  pid = NULL,
  uid_URL = TRUE,
  local_or_server = NULL,
  local_folder_tasks = NULL,
  server_folder_tasks = NULL,
  disable_web_security = FALSE,
  initial_wait = 2,
  wait_retry = 0.1,
  forced_random_wait = FALSE,
  forced_refresh = NULL,
  forced_seed = NULL,
  dont_ask = FALSE
)
}
\arguments{
\item{folder}{protocol folder}

\item{pid}{project id}

\item{dont_ask}{do everything without asking for user input}
}
\value{
Creates a _targets.R file
}
\description{
Create _targets.R file from a protocol folder
}
