% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{debug_docker}
\alias{debug_docker}
\title{debug_docker}
\usage{
debug_docker(uid_participant)
}
\arguments{
\item{uid_participant}{}
}
\description{
debug_docker
}
\examples{
debug_docker(24000)
}
