% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{list_data_server}
\alias{list_data_server}
\title{List files in the .data folder in the server}
\usage{
list_data_server(pid, list_credentials = NULL)
}
\arguments{
\item{pid}{protocol id}

\item{list_credentials}{list with credentials}
}
\description{
List files in the .data folder in the server
}
