% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{release_the_monkeys}
\alias{release_the_monkeys}
\title{release a horde of Monkeys to complete a jsPsychMakeR protocol}
\usage{
release_the_monkeys(
  uid = 1,
  credentials_folder = NULL,
  browserName = "chrome",
  big_container = FALSE,
  keep_alive = FALSE,
  folder_downloads = NULL,
  DEBUG = FALSE,
  screenshot = FALSE,
  debug_file = FALSE,
  console_logs = FALSE,
  open_VNC = FALSE,
  pid = NULL,
  uid_URL = TRUE,
  local_or_server = NULL,
  local_folder_tasks = NULL,
  server_folder_tasks = NULL,
  disable_web_security = FALSE,
  initial_wait = 2,
  wait_retry = 0.1,
  forced_random_wait = FALSE,
  forced_refresh = NULL,
  forced_seed = NULL,
  dont_ask = TRUE,
  open_rstudio = FALSE,
  clean_up_targets = FALSE
)
}
\arguments{
\item{open_rstudio}{}
}
\description{
release a horde of Monkeys to complete a jsPsychMakeR protocol
}
